Refactor vector store paths and improve embedding loading in quantization models

+++ quantization/model_4bit.py
@@ -11,7 +11,8 @@ from retrieval.database import Database
 MODEL_NAME = "jinaai/jina-embeddings-v2-base-code"
 EVAL_DATASET_PATH = "data/python_dataset/valid"
 TEST_DATASET_PATH = "data/python_dataset/test"
-TEMP_VECTORSTORE_PATH = "train/test_rag/temp_store"
+TEMP_VECTORSTORE_PATH_EVAL = "train/test_rag/temp_store_eval_4bit"
+TEMP_VECTORSTORE_PATH_TEST = "train/test_rag/temp_store_test_4bit"
 
 def eval():
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME, trust_remote_code=True)
@@ -41,6 +42,7 @@ def eval():
 
     # create a new vector-database for evaluation set
     # delete the temp vector store
+    TEMP_VECTORSTORE_PATH = TEMP_VECTORSTORE_PATH_EVAL
     if os.path.exists(TEMP_VECTORSTORE_PATH):
         for file_name in os.listdir(TEMP_VECTORSTORE_PATH):
             file_path = os.path.join(TEMP_VECTORSTORE_PATH, file_name)
@@ -82,24 +84,29 @@ def test():
     model_4bit = AutoModel.from_pretrained(MODEL_NAME, load_in_4bit=True, device_map="auto", trust_remote_code=True).to("cuda")
     model_4bit.eval()
     
-    test_dataset = PythonDataset(TEST_DATASET_PATH, tokenizer, max_len=512)
-    
-    for query_enc, code_enc in tqdm(test_dataset):
-        query_enc = {key: value.to("cuda") for key, value in query_enc.items()}
-        code_enc = {key: value.to("cuda") for key, value in code_enc.items()}
-        
+    if os.path.exists("test_embeddings_query_4bit.pt") and os.path.exists("test_embeddings_code_4bit.pt"):
+        queries = torch.load("test_embeddings_query_4bit.pt", weights_only=False)
+        codes = torch.load("test_embeddings_code_4bit.pt", weights_only=False)
+    else:
+        test_dataset = PythonDataset(TEST_DATASET_PATH, tokenizer, max_len=512)
         queries, codes = [], []
-        with torch.no_grad():
-            query_emb = model_4bit(**query_enc).pooler_output
-            code_emb = model_4bit(**code_enc).pooler_output
-            queries.append(query_emb)
-            codes.append(code_emb)
-    
-    # save the embeddings
-    torch.save(queries, "test_embeddings_query_4bit.pt")
-    torch.save(codes, "test_embeddings_code_4bit.pt")
+        for query_enc, code_enc in tqdm(test_dataset):
+            query_enc = {key: value.to("cuda") for key, value in query_enc.items()}
+            code_enc = {key: value.to("cuda") for key, value in code_enc.items()}
+            
+            with torch.no_grad():
+                query_emb = model_4bit(**query_enc).pooler_output
+                code_emb = model_4bit(**code_enc).pooler_output
+                queries.append(query_emb)
+                codes.append(code_emb)
+                assert type(query_emb) == torch.Tensor, "Query embeddings should be a PyTorch tensor"
+        
+        # save the embeddings
+        torch.save(queries, "test_embeddings_query_4bit.pt")
+        torch.save(codes, "test_embeddings_code_4bit.pt")
 
     # create a new vector-database for evaluation set
+    TEMP_VECTORSTORE_PATH = TEMP_VECTORSTORE_PATH_TEST
     if os.path.exists(TEMP_VECTORSTORE_PATH):
         for file_name in os.listdir(TEMP_VECTORSTORE_PATH):
             file_path = os.path.join(TEMP_VECTORSTORE_PATH, file_name)
