+++ quantization/model_8bit.py
@@ -12,8 +12,8 @@ from retrieval.database import Database
 MODEL_NAME = "jinaai/jina-embeddings-v2-base-code"
 EVAL_DATASET_PATH = "data/python_dataset/valid"
 TEST_DATASET_PATH = "data/python_dataset/test"
-TEMP_VECTORSTORE_PATH = "train/test_rag/temp_store"
-
+TEMP_VECTORSTORE_PATH_EVAL = "train/test_rag/temp_store_eval_8bit"
+TEMP_VECTORSTORE_PATH_TEST = "train/test_rag/temp_store_test_8bit"
 def eval():
     tokenizer = AutoTokenizer.from_pretrained(MODEL_NAME, trust_remote_code=True)
     model_8bit = AutoModel.from_pretrained(MODEL_NAME, load_in_8bit=True, device_map="auto", trust_remote_code=True).to("cuda")
@@ -42,6 +42,7 @@ def eval():
 
     # create a new vector-database for evaluation set
     # delete the temp vector store
+    TEMP_VECTORSTORE_PATH = TEMP_VECTORSTORE_PATH_EVAL
     if os.path.exists(TEMP_VECTORSTORE_PATH):
         for file_name in os.listdir(TEMP_VECTORSTORE_PATH):
             file_path = os.path.join(TEMP_VECTORSTORE_PATH, file_name)
@@ -83,24 +84,29 @@ def test():
     model_8bit = AutoModel.from_pretrained(MODEL_NAME, load_in_8bit=True, device_map="auto", trust_remote_code=True).to("cuda")
     model_8bit.eval()
     
-    test_dataset = PythonDataset(TEST_DATASET_PATH, tokenizer, max_len=512)
-    
-    for query_enc, code_enc in tqdm(test_dataset):
-        query_enc = {key: value.to("cuda") for key, value in query_enc.items()}
-        code_enc = {key: value.to("cuda") for key, value in code_enc.items()}
-        
+    if os.path.exists("test_embeddings_query_8bit.pt") and os.path.exists("test_embeddings_code_8bit.pt"):
+        queries = torch.load("test_embeddings_query_8bit.pt", weights_only=False)
+        codes = torch.load("test_embeddings_code_8bit.pt", weights_only=False)
+    else:
+        test_dataset = PythonDataset(TEST_DATASET_PATH, tokenizer, max_len=512)
         queries, codes = [], []
-        with torch.no_grad():
-            query_emb = model_8bit(**query_enc).pooler_output
-            code_emb = model_8bit(**code_enc).pooler_output
-            queries.append(query_emb)
-            codes.append(code_emb)
-    
-    # save the embeddings
-    torch.save(queries, "test_embeddings_query_8bit.pt")
-    torch.save(codes, "test_embeddings_code_8bit.pt")
+        for query_enc, code_enc in tqdm(test_dataset):
+            query_enc = {key: value.to("cuda") for key, value in query_enc.items()}
+            code_enc = {key: value.to("cuda") for key, value in code_enc.items()}
+            
+            with torch.no_grad():
+                query_emb = model_8bit(**query_enc).pooler_output
+                code_emb = model_8bit(**code_enc).pooler_output
+                queries.append(query_emb)
+                codes.append(code_emb)
+                assert type(query_emb) == torch.Tensor, "Query embeddings should be a PyTorch tensor"
+        
+        # save the embeddings
+        torch.save(queries, "test_embeddings_query_8bit.pt")
+        torch.save(codes, "test_embeddings_code_8bit.pt")
 
     # create a new vector-database for evaluation set
+    TEMP_VECTORSTORE_PATH = TEMP_VECTORSTORE_PATH_TEST
     if os.path.exists(TEMP_VECTORSTORE_PATH):
         for file_name in os.listdir(TEMP_VECTORSTORE_PATH):
             file_path = os.path.join(TEMP_VECTORSTORE_PATH, file_name)
@@ -111,12 +117,12 @@ def test():
         os.rmdir(TEMP_VECTORSTORE_PATH)
     
     db = Database(TEMP_VECTORSTORE_PATH)
-    for idx, (query_emb, code_emb) in tqdm(enumerate(zip(queries, codes)), total(len(queries)), desc="Loading embeddings"):
+    for idx, (query_emb, code_emb) in tqdm(enumerate(zip(queries, codes)), total=len(queries), desc="Loading embeddings"):
         metadata = {}
         code_id = f'code_{idx}'
         db.insert_embedding(code_emb.view(-1).tolist(), code_id, metadata)
     correct1, correctk, total = 0, 0, 0
-    for idx, (query_emb, code_emb) in tqdm(enumerate(zip(queries, codes)), total(len(queries)), desc="Evaluating embeddings"):
+    for idx, (query_emb, code_emb) in tqdm(enumerate(zip(queries, codes)), total=len(queries), desc="Evaluating embeddings"):
         query_embedding = query_emb.view(-1).tolist()
         retrievals = db.retrieve_by_embedding(query_embedding, top_k=5)
         retrieved_ids = retrievals['ids'][0]
